% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.cases.R
\docType{data}
\name{covid.cases}
\alias{covid.cases}
\title{Covid-19 Cases by Type (Confirmed, Recovered, Death)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2200 rows and 3 columns.
}
\source{
Epidemiological Unit, Ministry of Health, Sri Lanka \url{https://www.epid.gov.lk/web/}.
data(covid.cases)
}
\usage{
covid.cases
}
\description{
Daily summary of the Coronavirus (COVID-19) cases by state/province.
\describe{
\item{Date}{Date in YYYY-MM-DD format}
\item{Type}{An indicator for the type of cases (confirmed, death, recovered).}
\item{Cases}{Number of cases on given date.}}
}
\keyword{datasets}
