% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.wise.cases.R
\docType{data}
\name{district.wise.cases}
\alias{district.wise.cases}
\title{Covid-19 District-wise Confirmed Cases in Sri Lanka}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 832 rows and 3 columns.
}
\source{
Epidemiological Unit, Ministry of Health, Sri Lanka \url{https://www.epid.gov.lk/web/}.
data(district.wise.cases)
}
\usage{
district.wise.cases
}
\description{
Daily summary of the Coronavirus (COVID-19) cases by state/province.
\describe{
\item{Date}{Date in YYYY-MM-DD format}
\item{District}{Name of district.}
\item{Count}{Number of confirmed cases on given date.}}
}
\keyword{datasets}
