% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\docType{data}
\name{vaccination}
\alias{vaccination}
\title{Covid-19 Vaccination Counts in Sri Lanka}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 471 rows and 4 columns.
}
\source{
Epidemiological Unit, Ministry of Health, Sri Lanka \url{https://www.epid.gov.lk/web/}.
data(vaccination)
}
\usage{
vaccination
}
\description{
Daily summary of the Coronavirus (COVID-19) cases by state/province.
\describe{
\item{Date}{Date in YYYY-MM-DD format.}
\item{DVaccine}{Name of the vaccine.}
\item{\verb{First dose}}{Total number of people who have received the first dose on the given date.}
\item{\verb{Second dose}}{Total number of people who have received the first dose on the given date.}}
}
\keyword{datasets}
